[fix] migrate が実行できない

+++ db/migrate/20240503000032_fix32.rb
@@ -1,8 +1,9 @@
 class Fix32 < ActiveRecord::Migration[6.0]
   def up
-    tag = ActsAsTaggableOn::Tag.find_by(name: "対抗形")
-    if tag.taggings.count.zero?
-      tag.destroy!
+    if tag = ActsAsTaggableOn::Tag.find_by(name: "対抗形")
+      if tag.taggings.count.zero?
+        tag.destroy!
+      end
     end
 
     if tag = ActsAsTaggableOn::Tag.find_by(name: "対抗型")
